MIBE-59: chore: rename directory to fix naming issue
