+++ frontend/src/panel/components/pages/UserScriptSelectorPage.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useState } from 'react';
 import { useNavigationContext } from '../../contexts/contextHooks';
 import { getUser } from '../../api_temp/users';
-import APIResponse from '../../models/api/APIResponse';
+import APIResponse from '../../models/api_temp/APIResponse';
 import Loadable from '../Loadable';
 import List from '@mui/material/List/List';
 import ScriptListItem from '../../support_interface/script_selection/ScriptListItem';
@@ -13,7 +13,7 @@ import { createUnpublishedScript } from '../../api_temp/unpublishedScripts';
 import { editorScreen } from '../../navigation/screens';
 import Page from '../Page';
 import { useAPICall } from '../../api_temp/apiHooks';
-import { UserWithScripts } from '../../models/api/User';
+import { UserWithScripts } from '../../models/api_temp/User';
 
 const UserScriptSelectorPage = () => {
   const { goTo } = useNavigationContext();
